Add localized prefetching to support SPlaylistTracks and Add a fallback path to SPlaylistTracks for if the tracks paging does not exist
Localized prefetching runs before querying the FS cache, so can be used
to load SPlaylist for Cache.cachePlace (SPlaylistTracks _)

The fallback path is necessary in case SPlaylist is provided by
something other than a direct query. For example, SCurrentUserPlaylists
provides many of these, but does not provide the tracks paging to go
with them.