Add searching for artists, albums and tracks to WebClient and Query
Currently the queries only get the first page (50 items). This is
because it's not obvious how to provide access to the whole result. The
query can't just return all the results, because searches can return a
lot of results. Maybe have the query also take a page number and also
return whether there are more pages.